refactor(backend-jest-config): update backend jest config package with separated es modules
