+++ packages/backend-jest-config/lib/config/projectRoot.js
@@ -1,0 +1,3 @@
+const projectRoot = '<rootDir>';
+
+export default projectRoot;
