+++ packages/backend-jest-config/package.json
@@ -7,13 +7,17 @@
   "devDependencies": {
     "@jest/types": "29.3.1"
   },
+  "exports": "./lib/index.js",
   "homepage": "https://github.com/notaphplover/one-game#readme",
   "license": "See license in \"LICENSE\" file",
-  "main": "index.js",
   "name": "@one-game-js/backend-jest-config",
   "private": true,
   "repository": {
     "type": "git",
     "url": "git+https://github.com/notaphplover/one-game.git"
-  }
+  },
+  "scripts": {
+    "format": "prettier --write ./lib/**/*.js"
+  },
+  "type": "module"
 }
