+++ packages/backend-jest-config/prettier.config.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('@one-game-js/backend-prettier-config'),
+};
